// src/model/Product.ts
import mongoose, { Document, Schema, model } from "mongoose";

export interface ProductInterface extends Document {
  name: string;
  slug: string;
  image: string;
  brand: string;
  category: string;
  description: string;
  price: number;
  countInStock: number;
  rating: number;
  numReviews: number;
}

export interface Product {
  name: string;
  slug: string;
  image: string;
  brand: string;
  category: string;
  description: string;
  price: number;
  countInStock: number;
  rating: number;
  numReviews: number;
}

export const ProductSchema = new Schema<ProductInterface>({
  name: { type: String, required: true },
  slug: { type: String, required: true },
  image: { type: String, required: true },
  brand: { type: String, required: true },
  category: { type: String, required: true },
  description: { type: String, required: true },
  price: { type: Number, required: true, default: 0 },
  countInStock: { type: Number, required: true, default: 0 },
  rating: { type: Number, required: true, default: 0 },
  numReviews: { type: Number, required: true, default: 0 },
});

const Product = model<ProductInterface>("Products", ProductSchema);

export default Product;